Default to using fseeko()/ftello()/off_t on Linux.
The Linux build fails to define _LARGEFILE64_SOURCE macro, required to
make fseeko64() and ftello64() interfaces available when using GLIBC
(i.e. building with C compiler). This breaks support for large files.

Additionally, __off_t should not be used the offset type because it is
an internal type of GLIBC.

Because fseeko64() and others are intended only as interfaces for
transition to 64-bit offset it makes sense to witch to using ftello()
that provides 64-bit interface on both 64-bit and, with
_FILE_OFFSET_BITS=64 defined, 32-bit systems.