Fix failure to save mob from existing Structured Storage AAF file after changing its mob ID.
Fix OMStorable::detach() to use _exist, in addition to _store, to determine whether the OMStorable is present in the physical file. (_store is routinely cleared to reduce the count of open IStorage objects to accommodate limitations of Microsoft Structured Storage implementation.) This ensures that when an AAF object is detached from a file its stored representation, as well as those of its children (linked via strong references), are invalidated and can be properly re-created during save if the object is attached back.

This fixes a failure to save an existing mob after changing its ID with IAAFMob::SetMobID(), which works by removing and then adding the mob back to the file.